chore: consolidate file structure for platform adapter
- Rename `adapter/platform/grpc/BUILD.bazel` to `adapter/platform/BUILD.bazel`
- Rename `adapter/platform/grpc/impl.go` to `adapter/platform/impl.go`
- Rename `adapter/platform/wirex/injector.go` to `adapter/platform/injector.go`
- Rename `adapter/platform/grpc/main_test.go` to `adapter/platform/main_test.go`
- Rename `adapter/platform/grpc/router.go` to `adapter/platform/router.go`
- Rename `adapter/platform/grpc/web/static/css/app.css` to `adapter/platform/web/static/css/app.css`
- Rename `adapter/platform/grpc/web/static/image/auth0_logo_final_blue_RGB.png` to `adapter/platform/web/static/image/auth0_logo_final_blue_RGB.png`
- Rename `adapter/platform/grpc/web/static/js/js.cookie.js` to `adapter/platform/web/static/js/js.cookie.js`
- Rename `adapter/platform/grpc/web/static/js/user.js` to `adapter/platform/web/static/js/user.js`
- Rename `adapter/platform/grpc/web/templates/BUILD.bazel` to `adapter/platform/web/templates/BUILD.bazel`
- Rename `adapter/platform/grpc/web/templates/home.html` to `adapter/platform/web/templates/home.html`
- Rename `adapter/platform/grpc/web/templates/templates.go` to `adapter/platform/web/templates/templates.go`
- Rename `adapter/platform/grpc/web/templates/user.html` to `adapter/platform/web/templates/user.html`
- Rename `adapter/platform/grpc/wire.go` to `adapter/platform/wire.go`
- Rename `adapter
