+++ adapter/platform/web/static/css/app.css
