+++ adapter/platform/web/static/image/auth0_logo_final_blue_RGB.png
