+++ adapter/platform/web/static/js/js.cookie.js
