+++ adapter/platform/web/static/js/user.js
