--- adapter/platform/wirex/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "wirex",
-    srcs = ["injector.go"],
-    importpath = "github.com/blackhorseya/godine/adapter/platform/wirex",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//app/infra/authx",
-        "//app/infra/configx",
-    ],
-)
