+++ cmd/BUILD.bazel
@@ -10,21 +10,12 @@ go_library(
     importpath = "github.com/blackhorseya/godine/cmd",
     visibility = ["//visibility:public"],
     deps = [
-        "//adapter/logistics/restful",
-        "//adapter/notify/restful",
-        "//adapter/order/restful",
-        "//adapter/payment/restful",
-        "//adapter/platform/grpc",
-        "//adapter/restaurant/restful",
-        "//adapter/user/restful",
+        "//adapter/platform",
         "//app/infra/authz",
         "//app/infra/configx",
-        "//pkg/adapterx",
         "//pkg/cmdx",
-        "//pkg/contextx",
         "//pkg/logging",
         "@com_github_spf13_cobra//:cobra",
         "@com_github_spf13_viper//:viper",
-        "@org_golang_x_sync//errgroup",
     ],
 )
